+++ src/renderengine/DrawingTools/Dot.js
