+++ src/renderengine/DrawingTools/Fill.js
