+++ src/renderengine/DrawingTools/Obj.js
