+++ src/renderengine/DrawingTools/Panels.js
