+++ src/renderengine/DrawingTools/Polygon.js
