+++ src/renderengine/DrawingTools/Primitive.js
