+++ src/renderengine/DrawingTools/Rect.js
