+++ src/renderengine/DrawingTools/Stripes.js
