+++ src/renderengine/DrawingTools/createPixelArray.ts
